Separate TournamentBracket to a new file
